refactor mappers._hybrid module into _fine_res_budget, _fine_res, and _hybrid
