+++ external/loaders/loaders/mappers/__init__.py
@@ -10,9 +10,8 @@ from ._local import LocalMapper, mapper_to_local
 from ._hybrid import (
     open_fine_resolution_nudging_hybrid,
     open_precomputed_fine_resolution_nudging_hybrid,
-    open_fine_resolution,
-    open_precomputed_fine_resolution,
 )
+from ._fine_res import open_fine_resolution, open_precomputed_fine_resolution
 
 # additional open mapper functions
 from ._high_res_diags import open_high_res_diags
