+++ external/loaders/tests/test__fine_resolution_budget.py
@@ -1,7 +1,5 @@
 import numpy as np
-import xarray as xr
-from loaders.mappers._hybrid import _convergence, _extend_lower
-import pytest
+from loaders.mappers._fine_res_budget import _convergence
 
 
 def test__convergence_constant():
@@ -23,25 +21,3 @@ def test__convergence_linear():
 
     ans = _convergence(f, delp)
     np.testing.assert_almost_equal(ans, expected)
-
-
-@pytest.mark.parametrize(
-    ["nx", "nz", "vertical_dim", "error"],
-    (
-        pytest.param(1, 5, "z", False, id="base"),
-        pytest.param(5, 1, "x", False, id="different_dim"),
-        pytest.param(5, 1, "z", True, id="scalar_dim"),
-    ),
-)
-def test__extend_lower(nz, nx, vertical_dim, error):
-    f = xr.DataArray(np.arange(nx * nz).reshape((nx, nz)), dims=["x", "z"])
-    f.attrs = {"long_name": "some source term", "description": "some source term"}
-    ds = xr.Dataset({"f": f})
-    if error:
-        with pytest.raises(ValueError):
-            extended = _extend_lower(ds.f, vertical_dim)
-    else:
-        extended = _extend_lower(ds.f, vertical_dim)
-        xr.testing.assert_allclose(
-            extended.isel({vertical_dim: -1}), extended.isel({vertical_dim: -2})
-        )
